# SNOW-2105991: Use pre-computed row counts more aggressively