--- a/src/snowflake/snowpark/modin/plugin/_internal/frame.py
+++ b/src/snowflake/snowpark/modin/plugin/_internal/frame.py
@@ -738,10 +738,7 @@ class InternalFrame:
         Returns:
             Number of rows in this frame.
         """
-        num_rows = count_rows(self.ordered_dataframe)
-        self.ordered_dataframe.row_count = num_rows
-        self.ordered_dataframe.row_count_upper_bound = num_rows
-        return num_rows
+        return count_rows(self.ordered_dataframe)
 
     def has_unique_index(self, axis: Optional[int] = 0) -> bool:
         """