--- a/tests/integ/modin/frame/test_insert.py
+++ b/tests/integ/modin/frame/test_insert.py
@@ -54,7 +54,7 @@ def native_df():
         ),
     ],
 )
-@sql_count_checker(query_count=5, join_count=3)
+@sql_count_checker(query_count=4, join_count=3)
 def test_insert_snowpark_pandas_objects(native_df, native_value):
     snow_df = pd.DataFrame(native_df)
     value = pd.DataFrame(native_value)
@@ -99,7 +99,7 @@ def test_insert_snowpark_pandas_objects(native_df, native_value):
         ),
     ],
 )
-@sql_count_checker(query_count=2, join_count=1)
+@sql_count_checker(query_count=1, join_count=1)
 def test_insert_one_to_many(native_df, native_value):
     snow_df = pd.DataFrame(native_df)
     value = pd.DataFrame(native_value)
@@ -126,18 +126,18 @@ def test_insert_one_to_many(native_df, native_value):
 
 
 @pytest.mark.parametrize(
-    "value, expected_query_count, expected_join_count",
+    "value",
     [
-        (np.array(["a", "b", "c"]), 2, 1),  # numpy array of shape (N,)
-        (np.array([["a"], ["b"], ["c"]]), 2, 1),  # numpy array of shape (N, 1)
-        (["a", "b", "c"], 2, 1),  # python list
-        ({0: 1, 1: 2, 4: 3}, 1, 1),  # python dict
-        (("a", "b", "c"), 2, 1),  # python tuple
+        np.array(["a", "b", "c"]),  # numpy array of shape (N,)
+        np.array([["a"], ["b"], ["c"]]),  # numpy array of shape (N, 1)
+        ["a", "b", "c"],  # python list
+        {0: 1, 1: 2, 4: 3},  # python dict
+        ("a", "b", "c"),  # python tuple
     ],
 )
-def test_insert_array_like(native_df, value, expected_query_count, expected_join_count):
+def test_insert_array_like(native_df, value):
     snow_df = pd.DataFrame(native_df)
-    with SqlCounter(query_count=expected_query_count, join_count=expected_join_count):
+    with SqlCounter(query_count=1, join_count=1):
         eval_snowpark_pandas_result(
             snow_df,
             native_df,
@@ -187,7 +187,7 @@ def test_insert_pandas_types_negative(snow_df):
         snow_df.insert(0, "col3", value)
 
 
-@sql_count_checker(query_count=2)
+@sql_count_checker(query_count=1)
 def test_insert_dataframe_shape_negative(native_df):
     # DataFrame with more than one column
     snow_df = pd.DataFrame(native_df)
@@ -205,20 +205,20 @@ def test_insert_dataframe_shape_negative(native_df):
 
 
 @pytest.mark.parametrize(
-    "value, expected_query_count",
+    "value",
     [
         # NOTE: Accepted numpy array shapes are (N,) or (N, 1) where N = number of rows = 3
-        (np.ones((3, 2)), 0),
-        (np.ones((6, 1)), 1),
-        (np.ones((1, 1)), 1),
-        ([1, 2], 1),  # len < number of rows
-        ((6, 7, 8, 9), 1),  # len > number of rows
-        ({"a", "b", "c"}, 0),  # python set
+        np.ones((3, 2)),
+        np.ones((6, 1)),
+        np.ones((1, 1)),
+        [1, 2],  # len < number of rows
+        (6, 7, 8, 9),  # len > number of rows
+        {"a", "b", "c"},  # python set
     ],
 )
-def test_insert_value_negative(native_df, value, expected_query_count):
+def test_insert_value_negative(native_df, value):
     snow_df = pd.DataFrame(native_df)
-    with SqlCounter(query_count=expected_query_count):
+    with SqlCounter(query_count=0):
         eval_snowpark_pandas_result(
             snow_df,
             native_df,
@@ -227,7 +227,7 @@ def test_insert_value_negative(native_df, value, expected_query_count):
         )
 
 
-@sql_count_checker(query_count=2, join_count=1)
+@sql_count_checker(query_count=1, join_count=1)
 def test_insert_duplicate_label(native_df):
     snow_df = pd.DataFrame(native_df)
     eval_snowpark_pandas_result(
@@ -250,7 +250,7 @@ def test_insert_duplicate_label_negative(native_df):
 
 
 @pytest.mark.parametrize("loc", [0, 1, 2])
-@sql_count_checker(query_count=2, join_count=1)
+@sql_count_checker(query_count=1, join_count=1)
 def test_insert_loc(native_df, loc):
     snow_df = pd.DataFrame(native_df)
     eval_snowpark_pandas_result(
@@ -261,11 +261,9 @@ def test_insert_loc(native_df, loc):
     )
 
 
-@pytest.mark.parametrize(
-    "loc, expected_query_count", [(-99, 1), (-1, 1), (99, 1), ("1", 0)]
-)
-def test_insert_loc_negative(native_df, loc, expected_query_count):
-    with SqlCounter(query_count=expected_query_count):
+@pytest.mark.parametrize("loc", [-99, -1, 99, "1"])
+def test_insert_loc_negative(native_df, loc):
+    with SqlCounter(query_count=0):
         snow_df = pd.DataFrame(native_df)
         eval_snowpark_pandas_result(
             snow_df,
@@ -276,25 +274,23 @@ def test_insert_loc_negative(native_df, loc, expected_query_count):
 
 
 @pytest.mark.parametrize(
-    "value, expected_query_count, expected_join_count",
+    "value, expected_join_count",
     [
-        (np.array(["a", "b", "c", "d"]), 2, 1),  # numpy array of shape (N,)
-        (np.array([["a"], ["b"], ["c"], ["d"]]), 2, 1),  # numpy array of shape (N, 1)
-        (["a", "b", "c", "d"], 2, 1),  # python list
-        (("a", "b", "c", "d"), 2, 1),  # python tuple
-        ({(3, 1): 1}, 1, 1),  # python dict
-        ("abc", 1, 0),  # sting scalar
-        (1, 1, 0),  # int scalar
+        (np.array(["a", "b", "c", "d"]), 1),  # numpy array of shape (N,)
+        (np.array([["a"], ["b"], ["c"], ["d"]]), 1),  # numpy array of shape (N, 1)
+        (["a", "b", "c", "d"], 1),  # python list
+        (("a", "b", "c", "d"), 1),  # python tuple
+        ({(3, 1): 1}, 1),  # python dict
+        ("abc", 0),  # sting scalar
+        (1, 0),  # int scalar
     ],
 )
-def test_insert_multiindex_array_like_and_scalar(
-    value, expected_query_count, expected_join_count
-):
+def test_insert_multiindex_array_like_and_scalar(value, expected_join_count):
     arrays = [[3, 4, 5, 6], [1, 2, 1, 2]]
     index = pd.MultiIndex.from_arrays(arrays, names=["first", "second"])
     snow_df = pd.DataFrame({"col1": ["p", "q", "r", "s"]}, index=index)
     native_df = snow_df.to_pandas()
-    with SqlCounter(query_count=expected_query_count, join_count=expected_join_count):
+    with SqlCounter(query_count=1, join_count=expected_join_count):
         eval_snowpark_pandas_result(
             snow_df,
             native_df,
@@ -311,7 +307,7 @@ def test_insert_multiindex_array_like_and_scalar(
         ("a", "b", "c", "d"),  # python tuple
     ],
 )
-@sql_count_checker(query_count=2, join_count=1)
+@sql_count_checker(query_count=1, join_count=1)
 def test_insert_empty_multiindex_frame(value):
     mi = pd.MultiIndex.from_arrays([np.array([], dtype=int), np.array([], dtype=int)])
     snow_df = pd.DataFrame([], index=mi)
@@ -352,7 +348,7 @@ def test_insert_multiindex_dict_negative():
         ([1.0, 2.5, 3.0], [1, 2, 3]),  # Long and Double can be joined
     ],
 )
-@sql_count_checker(query_count=4, join_count=1)
+@sql_count_checker(query_count=3, join_count=1)
 def test_insert_compatible_index(df_index, value_index):
     snow_df = pd.DataFrame({"col1": ["p", "q", "r"]}, index=native_pd.Index(df_index))
     value = pd.DataFrame({"col2": ["x", "y", "z"]}, index=native_pd.Index(value_index))
@@ -408,7 +404,7 @@ def test_insert_index_num_levels_mismatch_negative(df_index, value_index):
         ),  # type mismatch boolean != long
     ],
 )
-@sql_count_checker(query_count=2, join_count=1)
+@sql_count_checker(query_count=1, join_count=1)
 def test_insert_index_type_mismatch(df_index, value_index, expected_index):
     # Note: This is different behavior than native pandas. In native pandas when
     # index datatype mismatch new columns in inserted will all NULL values.
@@ -464,16 +460,14 @@ def test_insert_multiple_null():
 
 
 @pytest.mark.parametrize(
-    "index, value, expected_query_count, expected_join_count",
+    "index, value",
     [
-        ([1, 2], native_pd.Series([1, 2], index=[2, 3]), 1, 1),
-        ([1, 2], [3, 4], 2, 1),
+        ([1, 2], native_pd.Series([1, 2], index=[2, 3])),
+        ([1, 2], [3, 4]),
     ],
 )
-def test_insert_into_empty_dataframe_with_index(
-    index, value, expected_query_count, expected_join_count
-):
-    with SqlCounter(query_count=expected_query_count, join_count=expected_join_count):
+def test_insert_into_empty_dataframe_with_index(index, value):
+    with SqlCounter(query_count=1, join_count=1):
         snow_df = pd.DataFrame(index=index)
         native_df = native_pd.DataFrame(index=index)
 
@@ -527,7 +521,7 @@ def test_insert_into_empty_dataframe(
     if isinstance(value, int):
         expected_join_count = 0
     if isinstance(value, list) or isinstance(value, np.ndarray):
-        expected_query_count = 2
+        expected_query_count = 1
     snow_df = pd.DataFrame(data=data, columns=columns)
     native_df = native_pd.DataFrame(data=data, columns=columns)
 
@@ -596,7 +590,7 @@ def test_insert_empty_list_into_empty_dataframe():
         ([], ["A", "B", "C"]),
     ],
 )
-@sql_count_checker(query_count=1)
+@sql_count_checker(query_count=0)
 def test_insert_into_empty_dataframe_negative(loc, data, columns):
     snow_df = pd.DataFrame(data=data, columns=columns)
     native_df = native_pd.DataFrame(data=data, columns=columns)
@@ -771,7 +765,7 @@ def test_insert_with_unique_and_duplicate_index_values(
         assert_frame_equal(snow_res, expected_res, check_dtype=False)
 
 
-@sql_count_checker(query_count=4, join_count=6)
+@sql_count_checker(query_count=3, join_count=6)
 def test_insert_timedelta():
     native_df = native_pd.DataFrame({"col1": [1, 2], "col2": [3, 4]})
     snow_df = pd.DataFrame(native_df)