<!---
Please answer these questions before creating your pull request. Thanks!
--->

1. Which Jira issue is this PR addressing? Make sure that there is an accompanying issue to your PR.

   <!---
   In this section, please add a Snowflake Jira issue number.

   Note that if a corresponding GitHub issue exists, you should still include
   the Snowflake Jira issue number. For example, for GitHub issue
   https://github.com/snowflakedb/snowpark-python/issues/1400, you should
   add "SNOW-1335071" here.
    --->

   Fixes SNOW-2105991

2. Fill out the following pre-review checklist:

   - [ ] I am adding a new automated test(s) to verify correctness of my new code
      - [ ] If this test skips Local Testing mode, I'm requesting review from @snowflakedb/local-testing
   - [ ] I am adding new logging messages
   - [ ] I am adding a new telemetry message
   - [ ] I am adding new credentials
   - [ ] I am adding a new dependency
   - [ ] If this is a new feature/behavior, I'm adding the Local Testing parity changes.
   - [ ] I acknowledge that I have ensured my changes to be thread-safe. Follow the link for more information: [Thread-safe Developer Guidelines](https://github.com/snowflakedb/snowpark-python/blob/main/CONTRIBUTING.md#thread-safe-development)
   - [ ] If adding any arguments to public Snowpark APIs or creating new public Snowpark APIs, I acknowledge that I have ensured my changes include AST support. Follow the link for more information: [AST Support Guidelines](https://github.com/snowflakedb/snowpark-python/blob/main/CONTRIBUTING.md#ast-abstract-syntax-tree-support-in-snowpark)

3. Please describe how your code solves the related issue.

SNOW-1900040 (#3144) added row count estimation that propagates across pandas dataframe operations, including storing precise row count information for frames constructed from native pandas/python objects and frames created by `read_snowflake`. This PR extends that earlier work by propagating the precise row count value across certain ordered dataframe operations, namely `select`, `union_all`, and `sort`, which all have predictable effects on the size of the resulting frame. This PR also changes internal methods that retrieve the frame's row count (used in many operations for bounds checking or other validation) to use this cached row count value rather than issuing a query.

In short, retrieving the length of a dataframe created from native pandas/python or directly by `read_snowflake` will no longer issue an extra query. This reduces query counts across large parts of the test suite. Some notable affected APIs include `repr`, `crosstab`, `insert`, `loc`, `iloc`, `iterrows`, and `groupby` operations with `by` specified as a native list object.

I did not take benchmarks for most of these operations; the removal of a query should strictly represent an improvement. I did benchmark changes to the `repr` operation, as some more work was needed to take advantage of the cached row count for that API (following the approach taken in SNOW-1705797/#2760). At some point between 4/15 and 4/22 ([dashboard metrics link](https://bench-dash.ordevmisc1.us-west-2.aws-dev.app.snowflake.com/benchmark/206296?metrics=261%2C262%2C266%2C267%2C272%2C276%2C277%2C841%2C842%2C846%2C847%2C851%2C852%2C856%2C857%2C861%2C862%2C1386%2C1387%2C1391%2C1392%2C1396%2C1397%2C1401%2C1402%2C1405%2C1406%2C1407%2C1411%2C1412%2C1416%2C1417%2C1421%2C1426%2C1427%2C1422%2C271&tab=chart); during this period the daily benchmark runner experienced some downtime due to modin versioning issues so the exact date is lost), repr for very large dataframes began taking almost twice as long as previously. I did not investigate the root cause, but this work remedies some of the performance impact.

### Performance for `repr(df)` on this PR (`b204452`) vs. main (`4b5feb`)
<table>
<tr><th></th><th>10 cols</th><th>100 cols</th><th>2K cols</th></tr>
<tr>
  <td>10 rows</td><td>0.534s -> 0.592s (+10.8%)</td>
  <td>0.648s -> 0.647s (-0.14%)</td>
  <td>19.3s -> 7.9s (-59.0%)</td>
</tr>
<tr>
  <td>1K rows</td>
  <td>0.408s -> 0.358s (-12.2%)</td>
  <td>0.681s -> 0.641s (-5.92%)</td>
  <td>23.1s -> 8.82s (-61.8%)</td>
</tr>
<tr>
  <td>1M rows</td>
  <td>0.979s -> 0.698s (-28.7%)</td>
  <td>2.17s -> 1.75s (-18.8%)</td>
  <td>53.6s -> 36.7s (-31.5%)</td>
</tr>
</table>