@sfc-gh-snowflakedb-snyk-sa 2025-05-13
![](https://res.cloudinary.com/snyk/image/upload/r-d/scm-platform/snyk-pull-requests/pr-banner-default.svg)

### :tada: **Snyk checks have passed. No issues have been found so far.**

:white_check_mark: **security/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/dc07b34e-1eae-4aab-acdb-7ee8df5efcd1?refs=ce39638170d03042657528bca786e265c0222b75&source=prChecksComment)

:white_check_mark: **license/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/dc07b34e-1eae-4aab-acdb-7ee8df5efcd1/license?refs=ce39638170d03042657528bca786e265c0222b75&source=prChecksComment)